+++ random-name-generator/plugins/random-name-generator/CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog random-name-generator
 
+## v1.0.2 - 2025-04-11
+* Better suggest name button for canary/tibiacom templates – like in the original game website
+
 ## v1.0.1 - 2025-11-25
-* Support for character name max-length attribute to the button
+* Support for the character name max-length attribute to the button
 * Fix for websites that use the 'account/character/create' link
 
 ## v1.0 - 2025-11-17